# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: nightshift
  version: 0.1.0
  description: >-
    Self-hosted task board where humans and AI agents are first-class actors. Plan A surface: tasks, tree/split,
    dependencies, claims with fencing tokens, audit, actor/token/policy administration. Threads/inbox/events arrive
    in later plans.
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
...
  /admin/actors:
    get:
      tags: [admin]
      operationId: listActors
      responses:
        "200":
          description: all actors
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Actor" } }
        default: { $ref: "#/components/responses/Problem" }
    post:
      tags: [admin]
      operationId: createActor
      responses:
        "201":
          description: created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Actor" }
        default: { $ref: "#/components/responses/Problem" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kind, handle, display_name]
              additionalProperties: false
              properties:
                kind: { type: string, enum: [human, agent] }
                handle: { type: string, minLength: 1, maxLength: 60 }
                display_name: { type: string, minLength: 1, maxLength: 120 }
                description: { type: string }
                # D-ss: humans only, optional; default member (agents: dropped by the use-case)
                role: { type: string, enum: [admin, member] }
  /admin/actors/{id}/tokens:
    parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
    post:
      tags: [admin]
      operationId: createToken
      description: raw_token shown exactly once; stored SHA-256 only (spec §5)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [label]
              additionalProperties: false
              properties: { label: { type: string, minLength: 1, maxLength: 60 } }
      responses:
        "201":
          description: issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  token_id: { type: string }
                  actor_id: { type: string }
                  label: { type: string }
                  created_at: { type: string }
                  raw_token: { type: string }
        default: { $ref: "#/components/responses/Problem" }
  /admin/tokens/{id}/revoke:
    parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
    post:
      tags: [admin]
      operationId: revokeToken
      responses:
        "204": { description: revoked }
        default: { $ref: "#/components/responses/Problem" }
  /admin/policy/{key}:
    parameters: [{ name: key, in: path, required: true, schema: { type: string } }]
    get:
      tags: [admin]
      operationId: getPolicy
      responses:
        "200":
          description: policy value (allowlisted keys)
          content:
            application/json:
              schema:
                type: object
                properties: { key: { type: string }, value: { type: ["string", "null"] } }
        default: { $ref: "#/components/responses/Problem" }
    put:
      tags: [admin]
      operationId: setPolicy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value]
              additionalProperties: false
              properties: { value: { type: string, enum: ["on", "off"] } }
      responses:
        "200":
          description: stored
          content:
            application/json:
              schema:
                type: object
                properties: { key: { type: string }, value: { type: string } }
        default: { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: agent tokens (and bootstrap admin) in Plan A; human cookie sessions arrive with OIDC in Plan E
...
  responses:
    Problem:
      description: domain rejection — stable machine code, branch on it (spec §11)
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    Actor:
      type: object
      properties:
        id: { type: string }
        kind: { type: string, enum: [human, agent] }
        # D-ss: admin|member for humans; null for agents (and pre-E legacy rows)
        role: { type: ["string", "null"], enum: ["admin", "member", null] }
        handle: { type: string }
        display_name: { type: string }
        description: { type: string }
        created_at: { type: string }
...
    Problem:
      type: object
      description: RFC 9457; `already_claimed` additionally carries holder_handle/holder_display_name
      required: [type, title, status, code]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        code:
          type: string
          enum:
            [
              not_found,
              forbidden,
              unauthenticated,
              invalid_request,
              internal_error,
              handle_taken,
              already_claimed,
              not_a_leaf,
              open_descendants,
              canceled_terminal,
              stale_lease,
              dependency_cycle,
              agent_close_forbidden,
              open_questions,
              threads_on_parent,
              question_transition,
              rate_limited,
              idempotency_in_flight,
              payload_too_large,
              unsupported_media_type,
            ]
        detail: { type: string }
...